got yarn workspaces working and added query and index and list to data adapter

+++ packages/adapter-fs/adapter.js
@@ -3,17 +3,39 @@ const path = require('path')
 const { Async } = require('crocks')
 const { always } = require('ramda')
 
-
+/**
+ * hyper63 adapter for the storage port
+ * 
+ * This adapter uses the file system
+ * as the implementation details for 
+ * the storage port.
+ *
+ *
+ * @typedef {Object} StorageInfo
+ * @property {string} bucket
+ * @property {string} object
+ * 
+ * @typedef {Object} StorageObject
+ * @property {string} bucket
+ * @property {string} object
+ * @property {stream} stream
+ *
+ * @typedef {Object} Response
+ * @property {boolean} ok
+ * @property {string} [msg] - error message
+ */
 /**
  * @param {string} path
  * @returns {Object}
  */
 module.exports = function (root) {
-
+  if (!root) { throw new Error('STORAGE: FS_Adapter: root directory required for this service!')}
   /**
    * @param {string} name
+   * @returns {Promise<Response>}
    */
   function makeBucket(name) {
+    if (!name) { return Promise.reject({ok: false, msg: 'name for bucket is required!'})}
     const mkdir = Async.fromNode(fs.mkdir)
     return mkdir(path.resolve(root + '/' + name))
       .map(always({ok: true}))
@@ -22,9 +44,10 @@ module.exports = function (root) {
 
   /**
    * @param {string} name
-   * 
+   * @returns {Promise<Response>} 
    */
   function removeBucket(name) {
+    if (!name) { return Promise.reject({ok: false, msg: 'name for bucket is required!'})}
     const rmdir = Async(function(reject, resolve) {
       fs.rmdir(path.resolve(root + '/' + name), (err) => {
         if (err) { return reject({ok: false, error: err.message })}
@@ -38,16 +61,14 @@ module.exports = function (root) {
   }
 
   /**
-   * @typedef {Object} StorageObject
-   * @property {string} bucket
-   * @property {string} object
-   * @property {stream} stream
-   */
-
-  /**
    * @param {StorageObject} 
+   * @returns {Promise<Response>}
    */
   function putObject ({bucket, object, stream}) {
+    if (!bucket) { return Promise.reject({ok: false, msg: 'bucket name required'})}
+    if (!object) { return Promise.reject({ok: false, msg: 'object name required'})}
+    if (!stream) { return Promise.reject({ok: false, msg: 'stream is required'})}
+
     return new Promise(function(resolve, reject) {
       const s = fs.createWriteStream(
         path.resolve(`${root}/${bucket}`) + `/${object}`  
@@ -58,7 +79,7 @@ module.exports = function (root) {
       })
 
       stream.on('error', (e) => {
-        reject({ok: false, error: e.message})
+        reject({ok: false, msg: e.message})
       })
 
       stream.pipe(s)
@@ -67,15 +88,19 @@ module.exports = function (root) {
   }
 
   /**
-   * @param {Object}
+   * @param {StorageInfo}
+   * @returns {Promise<Response>}
    */
   function removeObject({
     bucket,
     object
   }) {
+    if (!bucket) { return Promise.reject({ok: false, msg: 'bucket name required'})}
+    if (!object) { return Promise.reject({ok: false, msg: 'object name required'})}
+    
     const rm = Async(function(reject, resolve) {
       fs.unlink(path.resolve(`${root}/${bucket}/${object}`), (err) => {
-        if (err) { return reject({ok:false, error: err.message }) }
+        if (err) { return reject({ok:false, msg: err.message }) }
         resolve({ok: true})
       })
     })
@@ -83,28 +108,31 @@ module.exports = function (root) {
   }
 
   /**
-   * @param {Object}
+   * @param {StorageInfo}
+   * @returns {Promise<stream>}
    */
   function getObject({
     bucket, 
     object
   }) {
+    if (!bucket) { return Promise.reject({ok: false, msg: 'bucket name required'})}
+    if (!object) { return Promise.reject({ok: false, msg: 'object name required'})}
     return Async(function (reject, resolve) {
       try {
         let s = fs.createReadStream(path.resolve(`${root}/${bucket}/${object}`))
         resolve(s)
       } catch (e) {
-        reject({ok: false, error: e.message })
+        reject({ok: false, msg: e.message })
       }
     }).toPromise()
   }
 
-  function listObjects ({
-    bucket,
-    prefix = ""
-  }) {
+  function listObjects ({ bucket, prefix = "" }) {
+    if (!bucket) { return Promise.reject({ok: false, msg: 'bucket name required'})}
     return fs.promises.readdir(path.resolve(`${root}/${bucket}/${prefix}`))
   }
+
+
   return Object.freeze({
     makeBucket,
     removeBucket,
